EX-43 split code into two files. Test haven't been done yet
